Feature/userstudyroom

See merge request s09-webmobile1-sub2/S09P12A206!23

+++ backend/sswm/src/main/java/com/ground/sswm/userStudyroom/dto/UserStudyroomReqDto.java
@@ -11,7 +11,7 @@ public class UserStudyroomReqDto {
   private boolean isDeleted;
   private int totalStudy;
   private int totalRest;
-  private int userId;
-  private int studyroomId;
+  private Long userId;
+  private Long studyroomId;
 
 }
